Movie_Rating/requirements.txt
apache-airflow
pandas
psycopg2-binary
requests
